# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Layal Al Zahra Salon CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all clients
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Client list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientList"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a new client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get client by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /clients/{id}/history:
    get:
      operationId: getClientHistory
      tags: [clients]
      summary: Get appointment & purchase history for a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientHistory"

  # SERVICES
...
  /gift-card-types:
    get:
      operationId: listGiftCardTypes
      tags: [gift-cards]
      summary: List all gift card types
      responses:
        "200":
          description: Gift card types
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GiftCardType"
    post:
      operationId: createGiftCardType
      tags: [gift-cards]
      summary: Create a gift card type
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGiftCardTypeInput"
      responses:
        "201":
          description: Created type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GiftCardType"
  /gift-card-types/{id}:
    patch:
      operationId: updateGiftCardType
      tags: [gift-cards]
      summary: Update a gift card type
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGiftCardTypeInput"
      responses:
        "200":
          description: Updated type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GiftCardType"
    delete:
      operationId: deleteGiftCardType
      tags: [gift-cards]
      summary: Delete a gift card type
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
  /gift-cards:
    post:
      operationId: sellGiftCard
      tags: [gift-cards]
      summary: Sell a gift card to a client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SellGiftCardInput"
      responses:
        "201":
          description: Gift card sold
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GiftCardSale"
    get:
      operationId: listGiftCards
      tags: [gift-cards]
      summary: List all gift cards sold
      responses:
        "200":
          description: Gift card list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GiftCardSale"
...
  /clients/{id}/wallet:
    get:
      operationId: getClientWallet
      tags: [gift-cards]
      summary: Get client wallet balance and transactions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Wallet info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientWallet"
  /clients/{id}/wallet/redeem:
    post:
      operationId: redeemWallet
      tags: [gift-cards]
      summary: Deduct from client wallet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalletRedeemInput"
      responses:
        "200":
          description: Balance updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  clientId:
                    type: integer
                  balance:
                    type: number
components:
  schemas:
...
    Client:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        loyaltyPoints:
          type: integer
        totalSpent:
          type: number
        visitCount:
          type: integer
        lastVisit:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
    ClientInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        dateOfBirth:
          type: string
        gender:
          type: string
        notes:
          type: string
        tags:
          type: array
          items:
            type: string
    ClientUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        dateOfBirth:
          type: string
        gender:
          type: string
        notes:
          type: string
        tags:
          type: array
          items:
            type: string
    ClientList:
      type: object
      required: [clients, total, page, limit]
      properties:
        clients:
          type: array
          items:
            $ref: "#/components/schemas/Client"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    ClientHistory:
      type: object
      required: [appointments, sales]
      properties:
        appointments:
          type: array
          items:
            $ref: "#/components/schemas/Appointment"
        sales:
          type: array
          items:
            $ref: "#/components/schemas/Sale"

    # Services
...
    Appointment:
      type: object
      required: [id, clientId, clientName, staffId, staffName, serviceId, serviceName, date, startTime, endTime, status, totalPrice, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: string
        clientPhone:
          type: ["string", "null"]
        staffId:
          type: integer
        staffName:
          type: string
        serviceId:
          type: integer
        serviceName:
          type: string
        serviceCategoryName:
          type: ["string", "null"]
        serviceDuration:
          type: integer
        date:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        status:
          type: string
          enum: [pending, confirmed, in_progress, completed, cancelled, no_show]
        totalPrice:
          type: number
        notes:
          type: ["string", "null"]
        staffColor:
          type: ["string", "null"]
        paymentMethod:
          type: ["string", "null"]
        saleId:
          type: ["integer", "null"]
        createdAt:
          type: string
...
    SaleItem:
      type: object
      required: [id, type, name, quantity, unitPrice, totalPrice]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [service, product]
        referenceId:
          type: ["integer", "null"]
        name:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        totalPrice:
          type: number
    Sale:
      type: object
      required: [id, clientId, clientName, staffId, staffName, items, subtotal, discount, total, paymentMethod, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: ["integer", "null"]
        clientName:
          type: ["string", "null"]
        staffId:
          type: integer
        staffName:
          type: string
        appointmentId:
          type: ["integer", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItem"
        subtotal:
          type: number
        discount:
          type: number
        total:
          type: number
        paymentMethod:
          type: string
          enum: [cash, card, bank_transfer, loyalty_points]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
...
    GiftCardType:
      type: object
      required: [id, name, purchaseAmount, creditAmount, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        purchaseAmount:
          type: number
        creditAmount:
          type: number
        validityDays:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
    CreateGiftCardTypeInput:
      type: object
      required: [name, purchaseAmount, creditAmount]
      properties:
        name:
          type: string
        purchaseAmount:
          type: number
        creditAmount:
          type: number
        validityDays:
          type: integer
        isActive:
          type: boolean
    SellGiftCardInput:
      type: object
      required: [clientId, typeId]
      properties:
        clientId:
          type: integer
        typeId:
          type: integer
        paymentMethod:
          type: string
    GiftCardSale:
      type: object
      required: [id, clientId, clientName, typeId, typeName, referenceCode, purchaseAmount, creditAmount, remainingBalance, paymentMethod, status, purchasedAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: string
        typeId:
          type: integer
        typeName:
          type: string
        referenceCode:
          type: string
        purchaseAmount:
          type: number
        creditAmount:
          type: number
        remainingBalance:
          type: number
        paymentMethod:
          type: string
        expiryDate:
          type: string
        status:
          type: string
        purchasedAt:
          type: string
...
    WalletTransaction:
      type: object
      required: [id, type, amount, balanceAfter, description, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
        amount:
          type: number
        balanceAfter:
          type: number
        description:
          type: string
        referenceType:
          type: string
        referenceId:
          type: integer
        createdAt:
          type: string
    ClientWallet:
      type: object
      required: [clientId, balance, transactions]
      properties:
        clientId:
          type: integer
        balance:
          type: number
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/WalletTransaction"
    WalletRedeemInput:
      type: object
      required: [amount, description]
      properties:
        amount:
          type: number
        description:
          type: string
        referenceType:
          type: string
        referenceId:
          type: integer