# lib/api-zod/src/generated/types/giftCardSale.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Layal Al Zahra Salon CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface GiftCardSale {
  id: number;
  clientId: number;
  clientName: string;
  typeId: number;
  typeName: string;
  referenceCode: string;
  purchaseAmount: number;
  creditAmount: number;
  remainingBalance: number;
  paymentMethod: string;
  expiryDate?: string;
  status: string;
  purchasedAt: string;
}