# lib/api-zod/src/generated/types/giftCardType.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Layal Al Zahra Salon CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface GiftCardType {
  id: number;
  name: string;
  purchaseAmount: number;
  creditAmount: number;
  validityDays?: number;
  isActive: boolean;
  createdAt: string;
  updatedAt?: string;
}